draft new system to define emails
